Struct crushtool::CrushRuleMask

pub struct CrushRuleMask {
    pub ruleset: u8,
    pub rule_type: RuleType,
    pub min_size: u8,
    pub max_size: u8,
}

The rule mask is used to describe what the rule is intended for. Given a ruleset and size of output set, we search through the rule list for a matching rule_mask.

Fields
ruleset: u8 rule_type: RuleType min_size: u8 max_size: u8